Backported DefinitlyTyped changes
* Renamed pixi.d.ts to pixi.js.d.ts
* Extracted pixi-spine-tests.ts from pixi.d-tests.ts

Makes syncing with https://github.com/borisyankov/DefinitelyTyped easier

Bonus feature: `npm test` now runs the definition-tester using the locally installed typescript.
