+++ .gitignore
@@ -22,4 +22,4 @@ npm-debug.log
 !.gitkeep
 *__temp
 node_modules
-pixi-tests.js
+*-tests.js
