+++ README.md
@@ -2,6 +2,6 @@
 
 This definition file attempts to follow the [development branch of pixi](https://github.com/GoodBoyDigital/pixi.js/tree/dev) 
 
-The pixi-tests.ts attempts to keep up with the [latest examples](https://github.com/pixijs/examples)
+The pixi.js-tests.ts attempts to keep up with the [latest examples](https://github.com/pixijs/examples)
 
 Any advice, suggestions or improvements would be appreciated.
