+++ pixi.js.d.ts
